chore(typescript): migrate `victory-clip-container`
